update tests and ci
